Add files via upload
This dataset contains information on air traffic passenger statistics by the airline. It includes information on the airlines, airports, and regions that the flights departed from and arrived at. It also includes information on the type of activity, price category, terminal, boarding area, and number of passengers

Research Ideas

Air traffic passenger statistics could be used to predict future trends in air travel.

The data could be used to generate heat maps of airline traffic patterns.

The data could be used to study the effects of different factors on air traffic passenger numbers, such as the time of year or day, the price of airfare, or the number of flights offered by an airline